+++ doc/sphinx/source/releasenotes.md
@@ -9,6 +9,7 @@ On this page we provide a summary of the main API changes, new features and exam
 ### Interface changes
 
 - Add `bool` field type for `CeedQFunctionContext` and related interfaces to use `bool` fields.
+- `CEED_BASIS_COLLOCATED` removed; users should only use `CEED_BASIS_NONE`.
 
 ### New features